Introduce IndexDSL
The new index DSL solves some of the problems with the previous
stash_index approach to defining indexes:

stash_index, by default, automatically created indexes on all mentioned
fields.

It supported "only" and "except" to filter what kinds of indexes would
be applied.

When "unique" was added as an option, it exposed some awkwardness, e.g.

```
stash_index :name, :email, unique => true
```

would create all applicaable index types on name and email and add
*seperate* unique options for both of them (not a unique across two
fields).

The rules for what indexes and options were created also produced some
odd edge cases. A "match" index is not allowed to have "unique", so the
unique option would be ignored for those indexes.

Which led to the odd situation that

```
stash_index :name, except: [:range, :exact], unique: true
```

would not raise error (the "unique" option is useless in the above
example) but the following is a hard error:

```
stash_index :name, only: [:match], unique: true
```

would raise a hard error (even though semantically it is equivalent to
the above).

The new DSL tries to remain faithful to the intent of the previous DSL
while at the same time being easier to reason about. It also supports
indexing into associations.

Here is an example of the new DSL in action:

```
stash_index do
  # Automatically generate applicable indexes for the following fields
  # (exact, range and match)
  auto :first_name, :dob, :created_at

  # Create an exact index on these fields
  exact :gender, :email

  # Set the exact index on email to enforce uniqueness.
  # Note that if the exact index on email didn't already
  # exist, this would create one.
  unique :email

  # Create one match index from these three fields
  match_all :first_name, :last_name, :email

  # Index an association
  index_assoc :patient do
    range :height, :weight
  end
end
```